Crate cwe_checker_install

This crate automates the installation of cwe_checker. It creates config files, copies the Ghida-Plugin and can search for a Ghidra installation at commonly used locations.

Structs

CmdlineArgs

Installs cwe_checker

GhidraConfig

Structure for ghidra.json file

Functions

copy_config_json

Copies src/config.json to specified location

copy_dir_all

Recursive copy of files and directories.

copy_ghidra_plugin

Copy src/ghidra to provided location.

create_ghidra_json

Creates ghidra.json for a Ghidra location at provided locaton.

find_ghidra

Returns None if Ghidra were not found. Else returns path to Ghidra, which might be user selected.

get_search_locations

Returns vector of os-specific locations

get_user_choice

Determines Ghidra versions and provides selection interface for the user.

install_cwe_checker

Runs Cargo install to install cwe_checker.

is_good_ghidra_location

Determines if a path is a ghidra installation

main

search_for_ghidrarun

Searches for a file containing “ghidraRun” at provided path recursively.

select_ghidra_version

Determines Ghidra versions and provides selection interface for the user.

uninstall

Removes provided locations and uninstalls cwe_checker via cargo